Generate the USB, PCI, PNP and OUI vendor mapping from the hardware data
This allows us to match the vendor ID to a vendor name when we don't have
systemd providing a HWDB, for instance on Windows, ChromeOS, FreeBSD or Android.

Note: I actually tried building in the quirks into the quirks.xmlb cache, but
it increased the time to start the daemon by over 30x -- it turns out that it's
better to use a proper indexed database to look up a hashed key from tens of
thousands of results. Who knew!

The database cache is built at initial startup and is currently ~7MB in size.